Added inference notebook, train script and dataset
