+++ environment.yml
@@ -1,125 +1,16 @@
 name: text2sql
 channels:
-  - pytorch
-  - nvidia
   - conda-forge
   - defaults
 dependencies:
-  - blas=1.0=mkl
-  - brotli-python=1.1.0=py312h275cf98_2
-  - bzip2=1.0.8=h2466b09_7
-  - ca-certificates=2024.8.30=h56e8100_0
-  - certifi=2024.8.30=pyhd8ed1ab_0
-  - cffi=1.17.1=py312h4389bb4_0
-  - charset-normalizer=3.4.0=pyhd8ed1ab_0
-  - colorama=0.4.6=pyhd8ed1ab_0
-  - cuda-cccl=12.6.77=0
-  - cuda-cccl_win-64=12.6.77=0
-  - cuda-cudart=12.4.127=0
-  - cuda-cudart-dev=12.4.127=0
-  - cuda-cupti=12.4.127=0
-  - cuda-libraries=12.4.1=0
-  - cuda-libraries-dev=12.4.1=0
-  - cuda-nvrtc=12.4.127=0
-  - cuda-nvrtc-dev=12.4.127=0
-  - cuda-nvtx=12.4.127=0
-  - cuda-opencl=12.6.77=0
-  - cuda-opencl-dev=12.6.77=0
-  - cuda-profiler-api=12.6.77=0
-  - cuda-runtime=12.4.1=0
-  - cuda-version=12.6=3
-  - filelock=3.16.1=pyhd8ed1ab_0
-  - freetype=2.12.1=hdaf720e_2
-  - h2=4.1.0=pyhd8ed1ab_0
-  - hpack=4.0.0=pyh9f0ad1d_0
-  - hyperframe=6.0.1=pyhd8ed1ab_0
-  - idna=3.10=pyhd8ed1ab_0
-  - intel-openmp=2024.2.1=h57928b3_1083
-  - jinja2=3.1.4=pyhd8ed1ab_0
-  - lcms2=2.16=h67d730c_0
-  - lerc=4.0.0=h63175ca_0
-  - libblas=3.9.0=1_h8933c1f_netlib
-  - libcblas=3.9.0=5_hd5c7e75_netlib
-  - libcublas=12.4.5.8=0
-  - libcublas-dev=12.4.5.8=0
-  - libcufft=11.2.1.3=0
-  - libcufft-dev=11.2.1.3=0
-  - libcurand=10.3.7.77=0
-  - libcurand-dev=10.3.7.77=0
-  - libcusolver=11.6.1.9=0
-  - libcusolver-dev=11.6.1.9=0
-  - libcusparse=12.3.1.170=0
-  - libcusparse-dev=12.3.1.170=0
-  - libdeflate=1.22=h2466b09_0
-  - libexpat=2.6.3=he0c23c2_0
-  - libffi=3.4.2=h8ffe710_5
-  - libhwloc=2.11.1=default_h8125262_1000
-  - libiconv=1.17=hcfcfb64_2
-  - libjpeg-turbo=3.0.0=hcfcfb64_1
-  - liblapack=3.9.0=5_hd5c7e75_netlib
-  - libnpp=12.2.5.30=0
-  - libnpp-dev=12.2.5.30=0
-  - libnvfatbin=12.6.77=0
-  - libnvfatbin-dev=12.6.77=0
-  - libnvjitlink=12.4.127=0
-  - libnvjitlink-dev=12.4.127=0
-  - libnvjpeg=12.3.1.117=0
-  - libnvjpeg-dev=12.3.1.117=0
-  - libpng=1.6.44=h3ca93ac_0
-  - libsqlite=3.47.0=h2466b09_1
-  - libtiff=4.7.0=hfc51747_1
-  - libuv=1.49.2=h2466b09_0
-  - libwebp=1.4.0=h2466b09_0
-  - libwebp-base=1.4.0=hcfcfb64_0
-  - libxcb=1.16=h013a479_1
-  - libxml2=2.13.4=h442d1da_0
-  - libzlib=1.3.1=h2466b09_2
-  - m2w64-gcc-libgfortran=5.3.0=6
-  - m2w64-gcc-libs=5.3.0=7
-  - m2w64-gcc-libs-core=5.3.0=7
-  - m2w64-gmp=6.1.0=2
-  - m2w64-libwinpthread-git=5.0.0.4634.697f757=2
-  - markupsafe=3.0.2=py312h31fea79_0
-  - mkl=2023.1.0=h6a75c08_48682
-  - mpmath=1.3.0=pyhd8ed1ab_0
-  - msys2-conda-epoch=20160418=1
-  - networkx=3.4.2=pyhd8ed1ab_0
-  - numpy=2.1.2=py312hf10105a_0
-  - openjpeg=2.5.2=h3d672ee_0
-  - openssl=3.3.2=h2466b09_0
-  - pillow=10.4.0=py312h381445a_1
-  - pip=24.3.1=pyh8b19718_0
-  - pthread-stubs=0.4=hcd874cb_1001
-  - pthreads-win32=2.9.1=h2466b09_4
-  - pycparser=2.22=pyhd8ed1ab_0
-  - pysocks=1.7.1=pyh0701188_6
-  - python=3.12.4=h889d299_0_cpython
-  - python_abi=3.12=5_cp312
-  - pytorch=2.5.1=py3.12_cuda12.4_cudnn9_0
-  - pytorch-cuda=12.4=h3fd98bf_7
-  - pytorch-mutex=1.0=cuda
-  - pyyaml=6.0.2=py312h4389bb4_1
-  - requests=2.32.3=pyhd8ed1ab_0
-  - setuptools=75.1.0=pyhd8ed1ab_0
-  - sympy=1.13.3=pyh04b8f61_4
-  - tbb=2021.13.0=hc790b64_0
-  - tk=8.6.13=h5226925_1
-  - tqdm=4.66.6=pyhd8ed1ab_0
-  - typing_extensions=4.12.2=pyha770c72_0
-  - tzdata=2024b=hc8b5060_0
-  - ucrt=10.0.22621.0=h57928b3_1
-  - urllib3=2.2.3=pyhd8ed1ab_0
-  - vc=14.3=h8a93ad2_22
-  - vc14_runtime=14.40.33810=hcc2c482_22
-  - vs2015_runtime=14.40.33810=h3bf8584_22
-  - wheel=0.44.0=pyhd8ed1ab_0
-  - win_inet_pton=1.1.0=pyh7428d3b_7
-  - xorg-libxau=1.0.11=hcd874cb_0
-  - xorg-libxdmcp=1.1.3=hcd874cb_0
-  - xz=5.2.6=h8d14728_0
-  - yaml=0.2.5=h8ffe710_2
-  - zstandard=0.23.0=py312h7606c53_1
-  - zstd=1.5.6=h0ea2cb4_0
-  - pip:
-      - torchaudio==2.5.1
-      - torchvision==0.20.1
+  - python=3.12.4
+  - pytorch
+  - pytorch-cuda=12.4
+  - torchaudio
+  - torchvision
+  - tqdm
+  - conda-forge::outlines
+  - jupyter
+  - sqlparse
+  - accelerate
+  - conda-forge::records
