+++ src/lib/data.py
@@ -1,0 +1,10 @@
+from typing import List
+
+
+def prepare_input(question: str, table: List[str], tokenizer):
+    table_prefix = "table:"
+    question_prefix = "question:"
+    join_table = ",".join(table)
+    inputs = f"{question_prefix} {question} {table_prefix} {join_table}"
+    input_ids = tokenizer(inputs, max_length=700, return_tensors="pt").input_ids
+    return input_ids
